+++ jest.config.ts
@@ -6,5 +6,4 @@ export default {
   collectCoverage: true,
   coverageDirectory: 'coverage',
   coverageProvider: 'v8',
-  testMatch: ['**/tests/*.test.ts'],
 };